requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
greenlet==1.1.0
python-telegram-bot==13.5
pytz==2021.1
schedule==1.1.0
six==1.16.0
SQLAlchemy==1.4.16
tornado==6.1
tzlocal==2.1
